# PERF: faster categorical ops for equal or larger than scalar